chore: cleanup files that should not be edited
