+++ examples/react-ak-basic/src/config.ts
@@ -1,4 +1,3 @@
-import { createAppKit } from '@reown/appkit/basic'
 import { mainnet, solana } from '@reown/appkit/networks'
 import {
   useAppKit,
@@ -13,7 +12,21 @@ import {
 
 export const projectId = import.meta.env.VITE_PROJECT_ID || 'b56e18d47c72ab683b10814fe9495694' // this is a public projectId only to use on localhost
 
+// Create modal
+const modal = createAppKit({
+  networks: [mainnet, solana],
+  metadata: {
+    name: 'AppKit React Example',
+    description: 'AppKit React Ethers Example',
+    url: 'https://reown.com/appkit',
+    icons: ['https://avatars.githubusercontent.com/u/179229932?s=200&v=4']
+  },
+  projectId,
+  themeMode: 'light'
+})
+
 export {
+  modal,
   useAppKit,
   useAppKitState,
   useAppKitTheme,
