+++ examples/react-ak-basic/src/main.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client'
 
 import App from './App.jsx'
 import './assets/main.css'
-import './config.js'
 
 ReactDOM.createRoot(document.getElementById('app')!).render(
   <React.StrictMode>
