+++ examples/react-ep/vite.config.js
@@ -1,18 +1,6 @@
 import react from '@vitejs/plugin-react'
-import { defineConfig, loadEnv } from 'vite'
+import { defineConfig } from 'vite'
 
-// https://vite.dev/config/
-export default defineConfig(({ mode }) => {
-  const env = loadEnv(mode, process.cwd(), '')
-
-  return {
-    define: {
-      __APP_ENV__: JSON.stringify(env.APP_ENV),
-      'process.env': env
-    },
-    build: {
-      outDir: 'dist'
-    },
-    plugins: [react()]
-  }
+export default defineConfig({
+  plugins: [react()]
 })
