+++ packages/appkit-new/src/client.ts
@@ -848,10 +848,10 @@ export class AppKit {
             caipNetwork
           })
 
-          return result?.gas || BigInt(0)
+          return result?.gas || 0n
         }
 
-        return BigInt(0)
+        return 0n
       },
       getEnsAvatar: async () => {
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
@@ -893,7 +893,7 @@ export class AppKit {
       parseUnits: (value: string, decimals: number) => {
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
 
-        return adapter?.parseUnits({ value, decimals }) ?? BigInt(0)
+        return adapter?.parseUnits({ value, decimals }) ?? 0n
       },
       formatUnits: (value: bigint, decimals: number) => {
         const adapter = this.getAdapter(ChainController.state.activeChain as ChainNamespace)
