+++ packages/appkit/src/universal-adapter/client.ts
@@ -181,7 +181,7 @@ export class UniversalAdapter extends AdapterBlueprint {
   }
 
   public parseUnits(): AdapterBlueprint.ParseUnitsResult {
-    return BigInt(0)
+    return 0n
   }
 
   public formatUnits(): AdapterBlueprint.FormatUnitsResult {
