+++ packages/scaffold-ui/src/partials/w3m-connecting-wc-qrcode/index.ts
@@ -26,6 +26,7 @@ export class W3mConnectingWcQrcode extends W3mConnectingWidget {
   public constructor() {
     super()
     window.addEventListener('resize', this.forceUpdate)
+
     EventsController.sendEvent({
       type: 'track',
       event: 'SELECT_WALLET',
